+++ src/main/java/com/robertx22/age_of_exile/database/data/stats/types/defense/DamageShield.java
@@ -2,24 +2,20 @@ package com.robertx22.age_of_exile.database.data.stats.types.defense;
 
 import com.robertx22.age_of_exile.database.data.stats.Stat;
 import com.robertx22.age_of_exile.database.data.stats.StatScaling;
-import com.robertx22.age_of_exile.database.data.stats.effects.defense.DamageShieldEffect;
+import com.robertx22.age_of_exile.database.data.stats.effects.base.BaseDamageEffect;
+import com.robertx22.age_of_exile.saveclasses.unit.StatData;
+import com.robertx22.age_of_exile.uncommon.effectdatas.DamageEffect;
 import com.robertx22.age_of_exile.uncommon.enumclasses.Elements;
-import com.robertx22.age_of_exile.uncommon.interfaces.IStatEffect;
-import com.robertx22.age_of_exile.uncommon.interfaces.IStatEffects;
 
-public class DamageShield extends Stat implements IStatEffects {
+public class DamageShield extends Stat {
 
     private DamageShield() {
         this.scaling = StatScaling.NORMAL;
+        this.statEffect = new Effect();
     }
 
     public static DamageShield getInstance() {
         return SingletonHolder.INSTANCE;
-    }
-
-    @Override
-    public IStatEffect getEffect() {
-        return DamageShieldEffect.getInstance();
     }
 
     @Override
@@ -45,6 +41,32 @@ public class DamageShield extends Stat implements IStatEffects {
     @Override
     public String locNameForLangFile() {
         return "Damage Shield";
+    }
+
+    private static class Effect extends BaseDamageEffect {
+
+        @Override
+        public int GetPriority() {
+            return Priority.First.priority;
+        }
+
+        @Override
+        public EffectSides Side() {
+            return EffectSides.Target;
+        }
+
+        @Override
+        public DamageEffect activate(DamageEffect effect, StatData data, Stat stat) {
+            effect.number -= data.getAverageValue();
+
+            return effect;
+        }
+
+        @Override
+        public boolean canActivate(DamageEffect effect, StatData data, Stat stat) {
+            return true;
+        }
+
     }
 
     private static class SingletonHolder {