+++ src/main/java/com/robertx22/age_of_exile/database/data/stats/types/offense/NonCritDamage.java
@@ -2,16 +2,19 @@ package com.robertx22.age_of_exile.database.data.stats.types.offense;
 
 import com.robertx22.age_of_exile.database.data.stats.Stat;
 import com.robertx22.age_of_exile.database.data.stats.StatScaling;
-import com.robertx22.age_of_exile.database.data.stats.effects.offense.NonCritDamageEffect;
+import com.robertx22.age_of_exile.database.data.stats.effects.base.BaseDamageIncreaseEffect;
+import com.robertx22.age_of_exile.database.data.stats.effects.offense.crit.CriticalHitEffect;
+import com.robertx22.age_of_exile.saveclasses.unit.StatData;
+import com.robertx22.age_of_exile.uncommon.effectdatas.DamageEffect;
 import com.robertx22.age_of_exile.uncommon.enumclasses.Elements;
-import com.robertx22.age_of_exile.uncommon.interfaces.IStatEffect;
-import com.robertx22.age_of_exile.uncommon.interfaces.IStatEffects;
 
-public class NonCritDamage extends Stat implements IStatEffects {
+public class NonCritDamage extends Stat {
 
     private NonCritDamage() {
         this.scaling = StatScaling.SLOW;
         this.statGroup = StatGroup.MAIN;
+
+        this.statEffect = new Effect();
     }
 
     public static String GUID = "non_crit_damage";
@@ -41,13 +44,23 @@ public class NonCritDamage extends Stat implements IStatEffects {
     }
 
     @Override
-    public IStatEffect getEffect() {
-        return NonCritDamageEffect.getInstance();
-    }
-
-    @Override
     public String locNameForLangFile() {
         return "Non Critical Damage";
+    }
+
+    private static class Effect extends BaseDamageIncreaseEffect {
+
+        @Override
+        public int GetPriority() {
+            return Priority.afterThis(CriticalHitEffect.getInstance()
+                .GetPriority());
+        }
+
+        @Override
+        public boolean canActivate(DamageEffect effect, StatData data, Stat stat) {
+            return effect.isCriticalHit();
+        }
+
     }
 
     private static class SingletonHolder {