+++ src/main/java/com/robertx22/age_of_exile/database/data/stats/types/offense/ProjectileDamage.java
@@ -2,17 +2,19 @@ package com.robertx22.age_of_exile.database.data.stats.types.offense;
 
 import com.robertx22.age_of_exile.database.data.stats.Stat;
 import com.robertx22.age_of_exile.database.data.stats.StatScaling;
-import com.robertx22.age_of_exile.database.data.stats.effects.offense.damage_increase.ProjectileDamageEffect;
+import com.robertx22.age_of_exile.database.data.stats.effects.base.BaseDamageIncreaseEffect;
+import com.robertx22.age_of_exile.saveclasses.unit.StatData;
+import com.robertx22.age_of_exile.uncommon.effectdatas.DamageEffect;
+import com.robertx22.age_of_exile.uncommon.effectdatas.EffectUtils;
 import com.robertx22.age_of_exile.uncommon.enumclasses.Elements;
-import com.robertx22.age_of_exile.uncommon.interfaces.IStatEffect;
-import com.robertx22.age_of_exile.uncommon.interfaces.IStatEffects;
 
-public class ProjectileDamage extends Stat implements IStatEffects {
+public class ProjectileDamage extends Stat {
 
     public static String GUID = "projectile_damage";
 
     private ProjectileDamage() {
         this.scaling = StatScaling.NONE;
+        this.statEffect = new Effect();
     }
 
     public static ProjectileDamage getInstance() {
@@ -44,9 +46,13 @@ public class ProjectileDamage extends Stat implements IStatEffects {
         return "Projectile Damage";
     }
 
-    @Override
-    public IStatEffect getEffect() {
-        return ProjectileDamageEffect.getInstance();
+    private static class Effect extends BaseDamageIncreaseEffect {
+
+        @Override
+        public boolean canActivate(DamageEffect effect, StatData data, Stat stat) {
+            return EffectUtils.isProjectileAttack(effect);
+        }
+
     }
 
     private static class SingletonHolder {